moved the append_smiley onto restaurant into own file
