+++ src/database/models.rs
@@ -1,4 +1,5 @@
 use super::schema::*;
+use crate::database::append_smiley;
 use diesel::prelude::*;
 use serde::{Deserialize, Serialize};
 
@@ -31,7 +32,7 @@ impl Restaurant {
     pub fn testing_res_smiley(
         res_id: i32,
         conn: &SqliteConnection,
-    ) -> Vec<RestaurantWithSmileyReport> {
+    ) -> Vec<append_smiley::RestaurantWithSmileyReport> {
         use super::schema::restaurant::dsl::id;
         use super::schema::smiley_report::dsl::date;
         use super::schema::smiley_report::table as smiley_table;
@@ -43,7 +44,7 @@ impl Restaurant {
             .get_results::<(Restaurant, SmileyReport)>(conn)
             .ok()
             .expect("Error fetching restaurant ID");
-        convert_res_smiley_pairs(query)
+        append_smiley::convert_res_smiley_pairs(query)
     }
 
     pub fn get_restaurant_references(conn: &SqliteConnection) -> Vec<i32> {
@@ -135,6 +136,7 @@ impl Restaurant {
             .expect("Failed to get restaurants based on version")
     }
 }
+
 #[derive(Debug, Clone, PartialEq, Queryable, Serialize)]
 pub struct SmileyReport {
     pub id: i32,
@@ -156,75 +158,6 @@ impl SmileyReport {
         query.sort_by(|a, b| b.date.partial_cmp(&a.date).unwrap());
         query
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize)]
-pub struct RestaurantWithSmileyReport {
-    pub id: i32,
-    pub smiley_restaurant_id: i32,
-    pub name: String,
-    pub address: String,
-    pub zipcode: String,
-    pub city: String,
-    pub cvr: String,
-    pub pnr: String,
-    pub latitude: f32,
-    pub longitude: f32,
-    pub version_number: i32,
-    pub smileyreports: Vec<SmileyReport>,
-}
-
-fn convert_res_smiley_pairs(
-    input: Vec<(Restaurant, SmileyReport)>,
-) -> Vec<RestaurantWithSmileyReport> {
-    let mut result: Vec<RestaurantWithSmileyReport> = Vec::new();
-    let mut current: Restaurant = Restaurant {
-        id: -1,
-        smiley_restaurant_id: 0,
-        name: String::from(""),
-        address: String::from(""),
-        zipcode: String::from(""),
-        city: String::from(""),
-        cvr: String::from(""),
-        pnr: String::from(""),
-        latitude: 0.0,
-        longitude: 0.0,
-        version_number: 0,
-    };
-    let mut smileys: Vec<SmileyReport> = Vec::new();
-    for i in input {
-        if !(current.id == i.0.id) {
-            if !&smileys.is_empty() {
-                result.append(make_res_smiley(current, smileys).as_mut());
-                smileys = Vec::new();
-            }
-            println!("{:?}", &i.0);
-            current = i.0;
-        }
-        println!("{:?}", &i.1);
-        smileys.append(vec![i.1].as_mut());
-    }
-    result.append(make_res_smiley(current, smileys).as_mut());
-    result
-}
-fn make_res_smiley(
-    current: Restaurant,
-    smileys: Vec<SmileyReport>,
-) -> Vec<RestaurantWithSmileyReport> {
-    vec![RestaurantWithSmileyReport {
-        id: current.id,
-        smiley_restaurant_id: current.smiley_restaurant_id,
-        name: current.name,
-        address: current.address,
-        zipcode: current.zipcode,
-        city: current.city,
-        cvr: current.cvr,
-        pnr: current.pnr,
-        latitude: current.latitude,
-        longitude: current.longitude,
-        version_number: current.version_number,
-        smileyreports: smileys,
-    }]
 }
 
 use crate::services::subscription::SubscriptionRequest;
